Refactor model2vec integration: update dependencies, rename files, and enhance embedding registration
