+++ pyproject.toml
@@ -37,10 +37,10 @@ dependencies = [
 Homepage = "https://github.com/bhavnicksm/chonkie"
 
 [project.optional-dependencies]
-model2vec = ["sentence-transformers>=2.3.0", "numpy>=1.23.0", "model2vec"]
+model2vec = ["model2vec>=0.1.0", "numpy>=1.23.0"]
 semantic = ["sentence-transformers>=2.3.0", "numpy>=1.23.0"]
 openai = ["openai>=1.0.0", "numpy>=1.23.0"]
-all = ["sentence-transformers>=2.3.0", "numpy>=1.23.0", "openai>=1.0.0"]
+all = ["sentence-transformers>=2.3.0", "numpy>=1.23.0", "openai>=1.0.0", "model2vec>=0.1.0"]
 dev = ["pytest>=6.2.0", 
         "datasets>=1.14.0",
         "transformers>=4.0.0",
