+++ src/chonkie/embeddings/__init__.py
@@ -1,4 +1,5 @@
 from .base import BaseEmbeddings
+from .model2vec import Model2VecEmbeddings
 from .sentence_transformer import SentenceTransformerEmbeddings
 from .openai import OpenAIEmbeddings
 from .auto import AutoEmbeddings
@@ -7,6 +8,7 @@ from .auto import AutoEmbeddings
 # Add all embeddings classes to __all__
 __all__ = [
     "BaseEmbeddings",
+    "Model2VecEmbeddings", 
     "SentenceTransformerEmbeddings", 
     "OpenAIEmbeddings",
     "AutoEmbeddings",
